backend/pyproject.toml
[project]
name = "proposal-generator"
version = "0.1.0"
description = "AI-Powered Client Proposal Generator for 305-ai.com"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.42.0",
    "weasyprint>=62.0",
    "python-docx>=1.1.0",
    "docxtpl>=0.18.0",
    "cairosvg>=2.7.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.7.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
